--story=887060303 环境管理服务对接上云版作业平台
prefer：优化变量命名和类型

+++ src/backend/ci/core/environment/api-environment/src/main/kotlin/com/tencent/devops/environment/pojo/job/agentreq/QueryAgentStatusFromJobReq.kt
@@ -31,5 +31,5 @@ import io.swagger.annotations.ApiModelProperty
 
 data class QueryAgentStatusFromJobReq(
     @ApiModelProperty(value = "主机ID数组，单次查询主机数量不可超过5000")
-    val hostIdList: List<Int>
+    val hostIdList: List<Long>
 )
